+++ scratch/notes_john/cv.multinom-tests.R
@@ -21,5 +21,5 @@ m.beps <- update(m.beps, trace=FALSE)
 all.equal(cv1, cv2)
 
 system.time(print(cv1 <- cv(m.beps, k="loo")))
-system.time(print(cv2 <- cv(m.beps, k="loo", ncores=8)))
+system.time(print(cv2 <- cv(m.beps, k="loo", ncores=2)))
 all.equal(cv1, cv2)